Changed coord variables to doubles

+++ BubbleShooter/src/edu/sjsu/cs151/bubbleshooter/model/Bubble.java
@@ -15,8 +15,8 @@ public class Bubble
 	public boolean 	marked;
 	public double 		x;
 	public double 		y;
-	public int 		dx;
-	public int 		dy;
+	public double 		dx;
+	public double 		dy;
 	
 	public Bubble(Bubble thisLeft, Bubble thisRight, Bubble thisTopLeft, Bubble thisTopRight, Bubble thisBottomLeft, 
 	Bubble thisBottomRight)
